refactor(load-tests): move CLI into bin crate
Move the load test CLI to bin/load-tests and keep command logic in the load-tests library crate.
