+++ Justfile
@@ -11,7 +11,7 @@ mod actions 'actions'
 # Docker-based local devnet management
 mod devnet 'etc/docker'
 # Load testing for networks
-mod load-test 'crates/infra/load-tests'
+mod load-test 'etc/just/load-test.just'
 # Formatting, clippy, udeps, and deny checks
 mod check 'etc/just/check.just'
 # Cargo build targets and contract compilation
