+++ bin/load-tests/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "base-load-tests-bin"
+description = "Base Load Test Binary"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[[bin]]
+name = "base-load-tests"
+path = "src/main.rs"
+
+[lints]
+workspace = true
+
+[dependencies]
+url.workspace = true
+eyre.workspace = true
+base-load-tests.workspace = true
+alloy-signer-local.workspace = true
+clap = { workspace = true, features = ["derive", "env"] }
+tokio = { workspace = true, features = ["macros", "rt"] }
