+++ bin/load-tests/src/main.rs
@@ -1,0 +1,10 @@
+//! Load test runner binary.
+
+use clap::Parser;
+
+mod cli;
+
+#[tokio::main(flavor = "current_thread")]
+async fn main() -> eyre::Result<()> {
+    cli::Cli::parse().run().await
+}
