+++ crates/infra/load-tests/Cargo.toml
@@ -8,11 +8,6 @@ homepage.workspace = true
 repository.workspace = true
 description = "Core library for Base network load testing and benchmarking"
 
-[[bin]]
-name = "base-load-test"
-path = "src/bin/load_test.rs"
-
-
 [lints]
 workspace = true
 
@@ -53,8 +48,6 @@ tokio-util.workspace = true
 serde_json.workspace = true
 serde_yaml.workspace = true
 parking_lot.workspace = true
-reqwest = { workspace = true, features = ["json"] }
 tokio = { workspace = true, features = ["signal"] }
+reqwest = { workspace = true, features = ["json"] }
 tokio-tungstenite = { workspace = true, features = ["native-tls"] }
-
-[dev-dependencies]
