+++ crates/infra/load-tests/src/command/mod.rs
@@ -1,0 +1,7 @@
+//! Load-test command orchestration.
+
+mod load;
+pub use load::{LoadTest, LoadTestOptions};
+
+mod rescue;
+pub use rescue::{Rescue, RescueOptions};
