+++ crates/infra/load-tests/src/lib.rs
@@ -2,6 +2,9 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
+mod command;
+pub use command::{LoadTest, LoadTestOptions, Rescue, RescueOptions};
+
 mod config;
 pub use config::{
     OsakaTarget, PrecompileTarget, TestConfig, TxTypeConfig, WeightedTxType, WorkloadConfig,
