+++ etc/just/load-test.just
@@ -8,7 +8,7 @@
 #   FUNDER_KEY=0x... just load-test recover sepolia      - Recover funds from sepolia test accounts
 
 set positional-arguments := true
-set working-directory := '../../..'
+set working-directory := '../..'
 
 # Show load-test commands
 default:
@@ -16,17 +16,17 @@ default:
 
 # Run load test against devnet (local) - uses Anvil Account #1
 devnet *args:
-    FUNDER_KEY=0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d cargo run -p base-load-tests --bin base-load-test -- crates/infra/load-tests/examples/devnet.yaml {{args}}
+    FUNDER_KEY=0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d cargo run -p base-load-tests-bin -- crates/infra/load-tests/examples/devnet.yaml {{args}}
 
 # Run load test against devnet indefinitely (Ctrl-C to stop)
 devnet-continuous:
-    FUNDER_KEY=0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d cargo run -p base-load-tests --bin base-load-test -- crates/infra/load-tests/examples/devnet.yaml --continuous
+    FUNDER_KEY=0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d cargo run -p base-load-tests-bin -- crates/infra/load-tests/examples/devnet.yaml --continuous
 
 # Run load test against sepolia network (requires FUNDER_KEY env var)
 sepolia *args:
-    cargo run -p base-load-tests --bin base-load-test -- crates/infra/load-tests/examples/sepolia.yaml {{args}}
+    cargo run -p base-load-tests-bin -- crates/infra/load-tests/examples/sepolia.yaml {{args}}
 
 # Recover/drain funds from load test accounts back to funder (requires FUNDER_KEY env var)
 # Usage: just load-test recover <network>  (e.g., just load-test recover sepolia)
 recover network:
-    cargo run -p base-load-tests --bin base-load-test -- crates/infra/load-tests/examples/{{network}}.yaml --drain-only
+    cargo run -p base-load-tests-bin -- crates/infra/load-tests/examples/{{network}}.yaml --drain-only
